+++ cpu/op_rmw.h
@@ -86,7 +86,7 @@ CPU_OP_RMW_ADDR_W(trb)
 #define CPU_OP_RMW_ADDRX_B(op) \
    static inline void CPU_OP_RMW_ADDRX_B_DECL(op) (void) \
    { \
-      uint16_t addr = cpu_readw_pc() + REGS.x.w; \
+      uint32_t addr = (uint32_t)cpu_readw_pc() + (uint32_t)REGS.x.w; \
       uint8_t res = cpu_op_##op##_b (cpu_read(addr)); \
       cpu_write(addr, res); \
    }
@@ -96,7 +96,7 @@ CPU_OP_RMW_ADDR_W(trb)
 #define CPU_OP_RMW_ADDRX_W(op) \
    static inline void CPU_OP_RMW_ADDRX_W_DECL(op) (void) \
    { \
-      uint16_t addr = cpu_readw_pc() + REGS.x.w; \
+      uint32_t addr = (uint32_t)cpu_readw_pc() + (uint32_t)REGS.x.w; \
       uint16_t res = cpu_op_##op##_w (cpu_readw(addr)); \
       cpu_writew(addr, res); \
    }
